autodl/requirements.txt
huggingface-hub==0.30.2
transformers==4.51.3
fastapi==0.115.12
uvicorn==0.34.1
pydantic==2.11.3
unsloth==2025.3.19
python-dotenv==1.0.1